+++ stark/miden-lifted-stark/src/pcs/mod.rs
@@ -36,7 +36,5 @@ pub mod proof;
 pub mod prover;
 pub mod verifier;
 
-pub mod utils;
-
 #[cfg(test)]
 mod tests;